check count input
