+++ docs/README.md
@@ -9,7 +9,8 @@
 - [x] 자동차 이름 오류체크 하고 오류시 `alert`
 - [x] 횟수 입력창 띄우기
 - [x] 이동할 횟수 입력받기
-- [ ] 이동할 횟수 오류체크 하고 오류시 `alert`
+- [x] 이동할 횟수 오류체크 하고 오류시 `alert`
+- [x] 결과 표시 요소 띄우기
 - [ ] 자동차 객체 생성
 - [ ] random 값을 구해 이동을 결정하는 기능
 - [ ] 이동, 정지시 값 변화
