+++ src/elementControl.js
@@ -1,12 +1,19 @@
 export const showCountInput = () => {
   const racingCountDiv = document.querySelector('#racing-count-div');
-  const resultDiv = document.querySelector('#result');
-  const carNameSubmit = document.querySelector('#car-names-submit');
   const carNameInput = document.querySelector('#car-names-input');
+  const carNameSubmit = document.querySelector('#car-names-submit');
 
   racingCountDiv.style.display = 'block';
-  carNameSubmit.disabled = true;
   carNameInput.disabled = true;
+  carNameSubmit.disabled = true;
 };
 
-export const showResult = () => {};
+export const showResult = () => {
+  const resultDiv = document.querySelector('#result');
+  const racingCountInput = document.querySelector('#racing-count-input');
+  const racingCountSubmit = document.querySelector('#racing-count-submit');
+
+  resultDiv.style.display = 'block';
+  racingCountInput.disabled = true;
+  racingCountSubmit.disabled = true;
+};
